+++ OOP_Prcatice/java-lotto/src/main/java/lotto/utils/validators/WinLottoValidator.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.utils.validators;
 
 import lotto.model.Lotto;
