+++ OOP_Prcatice/java-lotto/src/test/java/lotto/utils/PurchaseValidatorTest.java
@@ -1,9 +1,8 @@
 package lotto.utils;
 
+import lotto.utils.validators.PurchaseValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
