+++ OOP_Prcatice/java-lotto/src/test/java/lotto/utils/WinLottoValidatorTest.java
@@ -1,5 +1,6 @@
 package lotto.utils;
 
+import lotto.utils.validators.WinLottoValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
